requirements.txt
ccxt~=1.73.85
pymongo~=4.0.1
cbpcommon~=0.0.10
dependency_injector~=4.39.1
PyYAML~=6.0
psycopg2~=2.9.3